Add: add ModelMapper e criação de crud de chamados

+++ src/main/java/br/com/mmt/helpdesk/domain/repository/ChamadoRepository.java
@@ -3,5 +3,5 @@ package br.com.mmt.helpdesk.domain.repository;
 import br.com.mmt.helpdesk.domain.Chamado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChamadoRepository extends JpaRepository<Chamado, Long> {
+public interface ChamadoRepository extends JpaRepository<Chamado, Integer> {
 }
